+++ example/actions/index.ts
@@ -1,5 +1,6 @@
 import { onChat } from "./chat";
 import { onPing } from "./ping";
 import { onSubscribe } from "./subscribe";
+import { onUnsubscribe } from "./unsubscribe";
 
-export const actions = [onChat, onPing, onSubscribe];
+export const actions = [onChat, onPing, onSubscribe, onUnsubscribe];