+++ src/remote-client.ts
@@ -1,20 +1,35 @@
-import { RemoteSocket } from "socket.io";
+import type { RemoteSocket } from "socket.io";
 import { z } from "zod";
-
 import { Distribution, makeDistribution } from "./distribution";
+import { EmissionMap, Emitter, EmitterConfig, makeEmitter } from "./emission";
 
-export interface RemoteClient<D extends z.SomeZodObject> extends Distribution {
+export type SomeRemoteSocket = RemoteSocket<
+  Record<string, (...args: any[]) => void>,
+  unknown
+>;
+
+export interface RemoteClient<E extends EmissionMap, D extends z.SomeZodObject>
+  extends Distribution {
   id: string;
   rooms: string[];
   getData: () => Readonly<Partial<z.infer<D>>>;
+  emit: Emitter<E>;
 }
 
-export const getRemoteClients = <D extends z.SomeZodObject>(
-  sockets: RemoteSocket<{}, z.infer<D>>[],
-) =>
-  sockets.map<RemoteClient<D>>((socket) => ({
+export const makeRemoteClients = <
+  E extends EmissionMap,
+  D extends z.SomeZodObject,
+>({
+  sockets,
+  ...rest
+}: {
+  sockets: SomeRemoteSocket[];
+  metadata: D;
+} & EmitterConfig<E>) =>
+  sockets.map<RemoteClient<E, D>>((socket) => ({
     id: socket.id,
     rooms: Array.from(socket.rooms),
-    getData: () => socket.data || {},
+    getData: () => (socket.data || {}) as Partial<D>,
+    emit: makeEmitter({ subject: socket, ...rest }),
     ...makeDistribution(socket),
   }));